+++ lib/Kconfig
@@ -1102,6 +1102,7 @@ config LMB
 	bool "Enable the logical memory blocks library (lmb)"
 	default y if ARC || ARM || M68K || MICROBLAZE || MIPS || \
 		     NIOS2 || PPC || RISCV || SANDBOX || SH || X86 || XTENSA
+	select ARCH_MISC_INIT if PPC
 	help
 	  Support the library logical memory blocks. This will require
 	  a malloc() implementation for defining the data structures
